Safari-609.1.11.1/LayoutTests/platform/mac-wk1/fast/dom/HTMLLinkElement/preconnect-support-expected.txt

Tests that Link's rel=preconnect is reported as supported.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testLink.relList.length is 1
PASS testLink.relList[0] is "preconnect"
FAIL testLink.relList.supports('preconnect') should be true. Was false.
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
